requirements.txt
opencv-python
pyserial
pillow
numpy 
matplotlib
